build fix
svn:sync-xref-src-repo-rev: r84823

+++ src/VBox/Runtime/common/path/RTPathParse.cpp.h
@@ -1,4 +1,4 @@
-/* $Id: RTPathParse.cpp.h 45394 2013-04-07 19:03:00Z knut.osmundsen@oracle.com $ */
+/* $Id: RTPathParse.cpp.h 45397 2013-04-08 08:37:31Z noreply@oracle.com $ */
 /** @file
  * IPRT - RTPathParse - Code Template.
  *
@@ -38,7 +38,7 @@ static int RTPATH_STYLE_FN(rtPathParse)(const char *pszPath, PRTPATHPARSED pPars
      * (keep it on the stack for speed).
      */
     uint32_t const  cMaxComps =  cbParsed < RT_UOFFSETOF(RTPATHPARSED, aComps[0xfff0])
-                              ? (cbParsed - RT_UOFFSETOF(RTPATHPARSED, aComps)) / sizeof(pParsed->aComps[0])
+                              ? (uint32_t)((cbParsed - RT_UOFFSETOF(RTPATHPARSED, aComps)) / sizeof(pParsed->aComps[0]))
                               : 0xfff0;
     uint32_t        idxComp   = 0;
     uint32_t        cchPath;
